[editor canvas] Add hooks directory. Factor out useCanvasStateChangeEffect.
